pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mysecond"
version = "0.1.0"
description = "Chess opening research tool – generates annotated PGN from engine + master-game data"
requires-python = ">=3.11"
dependencies = [
    "python-chess>=1.0",
    "click>=8.0",
    "requests>=2.28",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
mysecond = "mysecond.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mysecond"]
